make proxy send responses and add account-manager
