+++ packages/embark-accounts-manager/README.md
@@ -1,0 +1,6 @@
+# `embark-accounts-manager`
+
+Embark account manager. Takes the blockchain config and, along with the proxy, makes sure to send back the correct accounts
+
+Visit [embark.status.im](https://embark.status.im/) to get started with
+[Embark](https://github.com/embark-framework/embark).
