+++ packages/embark-accounts-manager/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "embark-proxy",
+  "name": "embark-accounts-manager",
   "version": "4.1.0-beta.5",
   "author": "Iuri Matias <iuri.matias@gmail.com>",
   "contributors": [],
-  "description": "Proxy to listen to blockchain requests and responses",
-  "homepage": "https://github.com/embark-framework/embark/tree/master/packages/embark-proxy#readme",
+  "description": "Embark account manager. Takes the blockchain config and, along with the proxy, makes sure to send back the correct accounts",
+  "homepage": "https://github.com/embark-framework/embark/tree/master/packages/embark-accounts-manager#readme",
   "bugs": "https://github.com/embark-framework/embark/issues",
   "keywords": [
     "blockchain",
@@ -20,7 +20,7 @@
   ],
   "license": "MIT",
   "repository": {
-    "directory": "packages/embark-proxy",
+    "directory": "packages/embark-accounts-manager",
     "type": "git",
     "url": "https://github.com/embark-framework/embark.git"
   },
@@ -45,17 +45,8 @@
   },
   "dependencies": {
     "@babel/runtime-corejs2": "7.3.1",
-    "embark-core": "^4.1.0-beta.5",
-    "embark-i18n": "^4.1.0-beta.3",
     "embark-utils": "^4.1.0-beta.5",
-    "ethereumjs-tx": "1.3.7",
-    "ethereumjs-util": "6.0.0",
-    "express": "4.17.1",
-    "http-proxy": "1.17.0",
-    "node-http-proxy-json": "0.1.6",
-    "pump": "3.0.0",
-    "simples": "0.8.8",
-    "stream-json": "1.1.3"
+    "web3": "1.0.0-beta.37"
   },
   "devDependencies": {
     "@babel/cli": "7.2.3",
