+++ packages/embark-accounts-manager/tsconfig.json
@@ -1,0 +1,5 @@
+{
+    "compilerOptions": { "baseUrl": ".", "paths": { "*": ["types/*"] } },
+  "extends": "../../tsconfig.json",
+  "include": ["src/**/*"]
+}
