+++ packages/embark/package.json
@@ -90,6 +90,7 @@
     "decompress": "4.2.0",
     "deep-equal": "1.0.1",
     "ejs": "2.6.1",
+    "embark-accounts-manager": "^4.1.0-beta.5",
     "embark-api": "^4.1.0-beta.5",
     "embark-authenticator": "^4.1.0-beta.5",
     "embark-blockchain-connector": "^4.1.0-beta.5",
